Admin
Crud Users amélioré
Gestion des rôles
Affichier roles ok
Creer role ok
ahouter role a un user en attente

+++ LoveMirorring/src/Api/Controllers/AdminController.cs
@@ -4,7 +4,9 @@ using System.Linq;
 using System.Threading.Tasks;
 using Api.Models;
 using Api.ViewModels.Admin;
+using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Http;
+using Microsoft.AspNetCore.Identity;
 using Microsoft.AspNetCore.Identity.UI.Services;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.EntityFrameworkCore;
@@ -14,96 +16,198 @@ namespace Api.Controllers
 {
     [Route("api/[controller]")]
     [ApiController]
+    //[Authorize(Roles = "Admin")]
     public class AdminController : ControllerBase
     {
         private readonly LoveMirroringContext _context;
         private readonly IEmailSender _emailSender;
+        private readonly RoleManager<AspNetRole> _roleManager;
+        private readonly UserManager<AspNetUser> _userManager;
 
-        public AdminController(LoveMirroringContext context, IEmailSender emailSender)
+
+        public AdminController(LoveMirroringContext context, IEmailSender emailSender, RoleManager<AspNetRole> roleManager, UserManager<AspNetUser> userManager)
         {
             _context = context;
             _emailSender = emailSender;
+            _roleManager = roleManager;
+            _userManager = userManager;
         }
 
+        //public AdminController(LoveMirroringContext context, IEmailSender emailSender)
+        //{
+        //    _context = context;
+        //    _emailSender = emailSender;
 
-        [Route("welcom")]
+        //}
+
+
+        [Route("Welcome")]
         [HttpGet]
         public async Task<IActionResult> Welcom()
         {
-            int accounts = await _context.AspNetUsers.CountAsync();
-            IndexModel overView = new IndexModel
+            try
             {
-                nbUsers = accounts
-            };
+                int accounts = await _context.AspNetUsers.CountAsync();
+                IndexModel overView = new IndexModel
+                {
+                    nbUsers = accounts
+                };
 
-            return new JsonResult(overView);
+                return new JsonResult(overView);
+            }
+            catch (Exception ex)
+            {
+                return StatusCode(500, ex);
+            }
+
+
         }
 
-        [Route("search/{username}")]
+        [Route("Search/{username}")]
         [HttpGet]
         public async Task<IActionResult> SearchUser(string username)
         {
-            string UserName = username.ToUpper();
+            try
+            {
+                if (username == null)
+                {
+                    throw new ArgumentNullException();
+                }
 
-            string id = (from u in await _context.AspNetUsers.ToListAsync()
-                         where u.NormalizedUserName.Equals(UserName)
-                         select u.Id).FirstOrDefault();
+                string UserName = username.ToUpper();
 
-            return new JsonResult(id);
+                string id = (from u in await _context.AspNetUsers.ToListAsync()
+                             where u.NormalizedUserName.Equals(UserName)
+                             select u.Id).FirstOrDefault();
+
+                if (id == null) 
+                {
+                    return new JsonResult(null);
+                }
+
+                return new JsonResult(id);
+            }
+            catch (ArgumentNullException)
+            {
+                return NotFound("Argument is null");
+            }
+            catch (Exception ex)
+            {
+                return StatusCode(500, ex);
+            }
+           
+
         }
 
-        [Route("user/{id}")]
+        [Route("GetUser/{id}")]
         [HttpGet]
         public async Task<IActionResult> GetUser(string id)
         {
-            SearchModel user = (from u in await _context.AspNetUsers.ToListAsync()
-                                where u.Id.Equals(id)
-                                select new SearchModel
-                                {
-                                    Id = u.Id,
-                                    UserName = u.UserName,
-                                    Email = u.Email,
-                                    EmailConfirmed = u.EmailConfirmed,
-                                    PhoneNumber = u.PhoneNumber,
-                                    PhoneNumberConfirmed = u.PhoneNumberConfirmed,
-                                    QuizCompleted = u.QuizCompleted
-                                }).FirstOrDefault();
+            try
+            {
+                if (id == null)
+                {
+                    throw new ArgumentNullException();
+                }
 
-            return new JsonResult(user);
+                SearchModel user = (from u in await _context.AspNetUsers.ToListAsync()
+                                    where u.Id.Equals(id)
+                                    select new SearchModel
+                                    {
+                                        Id = u.Id,
+                                        UserName = u.UserName,
+                                        Email = u.Email,
+                                        EmailConfirmed = u.EmailConfirmed,
+                                        PhoneNumber = u.PhoneNumber,
+                                        PhoneNumberConfirmed = u.PhoneNumberConfirmed,
+                                        QuizCompleted = u.QuizCompleted
+                                    }).FirstOrDefault();
+
+                if(user == null) 
+                {
+                    return new JsonResult(null);
+                }
+
+                return new JsonResult(user);
+            }
+            catch (ArgumentNullException)
+            {
+                return NotFound("Argument is null");
+            }
+            catch (Exception ex)
+            {
+                return StatusCode(500, ex);
+            }
+
         }
 
-        [Route("details/{id}")]
+        [Route("Details/{id}")]
         [HttpGet]
         public async Task<IActionResult> Details(string id)
         {
-            AspNetUser user = await _context.AspNetUsers.FindAsync(id);
+            try
+            {
+                if (id == null)
+                {
+                    throw new ArgumentNullException();
+                }
 
-            return new JsonResult(user);
+                AspNetUser user = await _context.AspNetUsers.FindAsync(id);
+
+                if (user == null)
+                {
+                    return new JsonResult(null);
+                }
+
+                return new JsonResult(user);
+            }
+            catch (ArgumentNullException)
+            {
+                return NotFound("Argument is null");
+            }
+            catch (Exception ex)
+            {
+                return StatusCode(500, ex);
+            }
+
         }
 
-        [Route("edit")]
+        [Route("Edit")]
         [HttpPut("{id}")]
         public async Task<IActionResult> Edit(AspNetUser user)
         {
-            _context.Entry(user).State = EntityState.Modified;
-
             try
             {
+                if (user == null)
+                {
+                    throw new ArgumentNullException();
+                }
+
+                _context.Entry(user).State = EntityState.Modified;
                 await _context.SaveChangesAsync();
+
+                return NoContent();
             }
+            catch (ArgumentNullException)
+            {
+                return NotFound("Argument is null");
+            }
+
             catch (DbUpdateConcurrencyException)
             {
                 if (!UserExists(user.Id))
                 {
-                    return NotFound();
+                    return new JsonResult(null);
                 }
                 else
                 {
                     throw;
                 }
             }
-
-            return NoContent();
+            catch (Exception ex)
+            {
+                return StatusCode(500, ex);
+            }
 
         }
 
@@ -112,20 +216,88 @@ namespace Api.Controllers
             return _context.AspNetUsers.Any(e => e.Id.Equals(id));
         }
 
-        [Route("delete/{id}")]
+        [Route("Delete/{id}")]
         [HttpDelete]
         public async Task<IActionResult> Delete(string id)
         {
-            AspNetUser user = await _context.AspNetUsers.FindAsync(id);
-            _context.AspNetUsers.Remove(user);
-            await _context.SaveChangesAsync();
+            try
+            {
+                if (id == null)
+                {
+                    throw new ArgumentNullException();
+                }
 
-            await _emailSender.SendEmailAsync(
-         user.Email,
-         "Your account has been deleted by the Administrator",
-         "Your account has been deleted</br></br> Have a nice day !");
+                AspNetUser user = await _context.AspNetUsers.FindAsync(id);
 
-            return Ok();
+                if (user == null)
+                {
+                    return new JsonResult(null);
+                }
+
+                _context.AspNetUsers.Remove(user);
+                await _context.SaveChangesAsync();
+
+                await _emailSender.SendEmailAsync(
+             user.Email,
+             "Your account has been deleted by the Administrator",
+             "Your account has been deleted</br></br> Have a nice day !");
+
+                return Ok();
+
+            }
+            catch (ArgumentNullException)
+            {
+                return NotFound("Argument is null");
+            }
+            catch (Exception ex)
+            {
+                return StatusCode(500, ex);
+            }
+
+        }
+
+        [Route("Roles")]
+        [HttpGet]
+        public async Task<IActionResult> UserRole()
+        {
+            var roles = await _context.AspNetRoles.ToListAsync();
+            var users = await _context.AspNetUsers.ToListAsync();
+            var userRoles = from u in await _context.AspNetUsers.ToListAsync()
+                            join ur in await _context.AspNetUserRoles.ToListAsync() on u.Id equals ur.UserId
+                            join r in await _context.AspNetRoles.ToListAsync() on ur.RoleId equals r.Id
+                            select r.Name;
+
+            return new JsonResult(new RolesModel
+            {
+                Roles = roles.Select(x => x.NormalizedName),
+                Users = users.Select(u => new UsersModel {Email = u.Email, UserId = u.Id, Roles = userRoles  })
+            });
+        }
+
+        [Route("CreateRole")]
+        [HttpPost]
+        public async Task<IActionResult> CreateRole(AspNetRole role)
+        {
+            await _roleManager.CreateAsync(new AspNetRole { Name = role.Name, NormalizedName = role.Name.ToUpper() });
+
+            return NoContent();
+        }
+
+        [Route("UpdateUserRole")]
+        [HttpPost]
+        public async Task<IActionResult> UpdateUserRole(UpdateUserRoleModel vm)
+        {
+            if (vm.UserEmail != null && vm.Role != null)
+            {
+
+                var user = await _userManager.FindByEmailAsync(vm.UserEmail);
+
+                if (vm.Delete)
+                    await _userManager.RemoveFromRoleAsync(user, vm.Role);
+                else
+                    await _userManager.AddToRoleAsync(user, vm.Role);
+            }
+            return NoContent();
         }
 
     }
